simplify import with confirm dialog
