--- app/src/main/res/drawable/ic_photo_gallery.xml
@@ -1,34 +1,0 @@
-<!--
-  ~   Copyright 2020-2021 Leon Latsch
-  ~
-  ~   Licensed under the Apache License, Version 2.0 (the "License");
-  ~   you may not use this file except in compliance with the License.
-  ~   You may obtain a copy of the License at
-  ~
-  ~        http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~   Unless required by applicable law or agreed to in writing, software
-  ~   distributed under the License is distributed on an "AS IS" BASIS,
-  ~   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~   See the License for the specific language governing permissions and
-  ~   limitations under the License.
-  -->
-
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="410.309dp"
-    android:height="410.309dp"
-    android:viewportWidth="410.309"
-    android:viewportHeight="410.309">
-    <path
-        android:fillColor="#00ACEA"
-        android:pathData="M382.955,58.96c16.936,2.176 29.014,17.507 27.167,34.482L386.09,306.079c-1.848,16.923 -17.066,29.144 -33.989,27.295c-0.339,-0.037 -0.677,-0.08 -1.015,-0.128h-1.567V138.372c0,-17.312 -14.035,-31.347 -31.347,-31.347H56.947l5.747,-52.245c2.179,-17.223 17.742,-29.535 35.004,-27.69L382.955,58.96z" />
-    <path
-        android:fillColor="#00CEB4"
-        android:pathData="M349.518,333.246v18.808c0,17.312 -14.035,31.347 -31.347,31.347H31.347C14.035,383.401 0,369.366 0,352.054v-43.886l86.204,-62.694c13.668,-10.37 32.794,-9.491 45.453,2.09l57.469,50.155c12.046,10.215 29.238,11.683 42.841,3.657l117.551,-68.963V333.246z" />
-    <path
-        android:fillColor="#00EFD1"
-        android:pathData="M349.518,138.372v94.041l-117.551,68.963c-13.603,8.026 -30.795,6.558 -42.841,-3.657l-57.469,-50.155c-12.659,-11.58 -31.785,-12.46 -45.453,-2.09L0,308.168V138.372c0,-17.312 14.035,-31.347 31.347,-31.347h286.824C335.484,107.026 349.518,121.06 349.518,138.372z" />
-    <path
-        android:fillColor="#D4E1F4"
-        android:pathData="M208.98,192.707m-33.437,0a33.437,33.437 0,1 1,66.874 0a33.437,33.437 0,1 1,-66.874 0" />
-</vector>
