--- app/src/main/res/layout/dialog_receive_share.xml
@@ -1,79 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~   Copyright 2020-2021 Leon Latsch
-  ~
-  ~   Licensed under the Apache License, Version 2.0 (the "License");
-  ~   you may not use this file except in compliance with the License.
-  ~   You may obtain a copy of the License at
-  ~
-  ~        http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~   Unless required by applicable law or agreed to in writing, software
-  ~   distributed under the License is distributed on an "AS IS" BASIS,
-  ~   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~   See the License for the specific language governing permissions and
-  ~   limitations under the License.
-  -->
-
-<layout xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <data>
-
-        <variable
-            name="context"
-            type="dev.leonlatsch.photok.ui.share.ReceiveShareDialog" />
-
-        <variable
-            name="viewModel"
-            type="dev.leonlatsch.photok.ui.share.ReceiveShareViewModel" />
-
-        <import type="dev.leonlatsch.photok.databinding.BindingConverters" />
-    </data>
-
-    <LinearLayout
-        style="@style/LinearDialogStyle"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <TextView
-            style="@style/BigTextViewStyle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="15dp"
-            android:text="Import Photos" />
-
-        <ImageView
-            android:layout_width="wrap_content"
-            android:layout_height="200dp"
-            android:src="@drawable/ic_photo_gallery" />
-
-        <GridLayout
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content">
-
-            <TextView
-                style="@style/LabelTextViewStyle"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_row="0"
-                android:layout_column="0"
-                android:text="Elements:" />
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_row="0"
-                android:layout_column="1"
-                android:text="@{BindingConverters.INSTANCE.toStringConverter(viewModel.elementsToProcess)}" />
-
-        </GridLayout>
-
-        <Button
-            android:id="@+id/changePasswordCheckOldButton"
-            android:layout_width="@dimen/wide_button_width"
-            android:layout_height="@dimen/button_height"
-            android:layout_marginTop="@dimen/medium_gap"
-            android:onClick="@{() -> context.startImport()}"
-            android:text="Import" />
-
-    </LinearLayout>
-</layout>
