+++ app/src/main/res/values-pt-rBR/strings.xml
@@ -35,6 +35,7 @@
     <!-- IMPORT -->>
     <string name="import_importing">Importando foto(s)…</string>
     <string name="import_permission_rationale">Photok precisa acessar seu armazenamento, para importar fotos.</string>
+    <string name="import_sharted_question">Do you want to import %1$d photo(s)?</string><!-- TODO: Translate -->
 
     <!-- SETUP -->
     <string name="setup_create_your_password">Criar sua senha</string>
