+++ app/src/main/res/values-zh-rCN/strings.xml
@@ -35,6 +35,7 @@
     <!-- IMPORT -->>
     <string name="import_importing">正在导入照片……</string>
     <string name="import_permission_rationale">Photok 需要访问你的存储来导入照片</string>
+    <string name="import_sharted_question">Do you want to import %1$d photo(s)?</string><!-- TODO: Translate -->
 
     <!-- SETUP -->
     <string name="setup_create_your_password">创建你的密码</string>
