+++ app/src/main/res/values/strings.xml
@@ -34,6 +34,7 @@
     <!-- IMPORT -->>
     <string name="import_importing">Importing Photo(s)…</string>
     <string name="import_permission_rationale">Photok needs to access your storage, to import photos.</string>
+    <string name="import_sharted_question">Do you want to import %1$d photo(s)?</string>
 
     <!-- SETUP -->
     <string name="setup_create_your_password">Create your password</string>
