+++ icon_credits.html
@@ -17,8 +17,6 @@
 <h3>Credits for Icons:</h3>
 <div>
     Icons made by <a href="https://www.flaticon.com/authors/google" title="Google">Google</a>,
-    <a href="https://www.flaticon.com/free-icon/gallery_833453?related_item_id=833281&term=photo"
-       title="Kiranshastry">Kiranshastry</a>,
     <a href="https://www.flaticon.com/authors/freepik" title="Freepik">Freepik</a> and
     <a href="https://www.flaticon.com/authors/ultimatearm" title="ultimatearm">ultimatearm</a> from
     <a href="https://www.flaticon.com/" title="Flaticon">www.flaticon.com</a>
